+++ test/test_helper.rb
@@ -4,4 +4,5 @@ $LOAD_PATH.unshift File.expand_path("../lib", __dir__)
 
 require "minitest/autorun"
 require "minitest/spec"
+require "async"
 require "nnq/zstd"